category product links

+++ app/Http/Controllers/CategoryController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Category;
+use App\User;
 
 class CategoryController extends Controller
 {
@@ -15,6 +17,10 @@ class CategoryController extends Controller
 
     public function index()
     {
+        if (!$this->isAdmin()) {
+            return redirect('home');
+        }
+
         $categories = Category::all();
 
         return view('category/category', ['categories' => $categories]);
@@ -22,11 +28,20 @@ class CategoryController extends Controller
 
     public function add()
     {
+        if (!$this->isAdmin()) {
+            return redirect('home');
+        }
+
         return view("category/add");
     }
 
     public function store(Request $request)
     {
+        $request->validate([
+            'title' => ['required'],
+            'image' => ['required'],
+        ]);
+
         $categorie = new Category;
 
         $categorie->title = $request->input('title');
@@ -34,15 +49,26 @@ class CategoryController extends Controller
 
         $categorie->save();
         
-        return back()->withInfo("Category add");
+        return redirect('categories')->withInfo('Category add');
     }
 
     public function delete($id)
     {
+        if (!$this->isAdmin()) {
+            return redirect('home');
+        }
+
         $categorie = Category::find($id);
         $categorie->delete();
 
         return back()->withInfo("Category delete");
+    }
+
+    private function isAdmin()
+    {
+        $user = Auth::user();
+
+        return (Auth::check() && $user->role === "admin");
     }
 
 }
