+++ app/Http/Controllers/ProductController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Product;
 use App\Category;
+use App\User;
 
 class ProductController extends Controller
 {
@@ -15,6 +17,10 @@ class ProductController extends Controller
 
     public function index()
     {
+        if (!$this->isAdmin()) {
+            return redirect('home');
+        }
+
         $products = Product::all();
 
         return view('product/product', ['products' => $products]);
@@ -22,6 +28,9 @@ class ProductController extends Controller
 
     public function add()
     {
+        if (!$this->isAdmin()) {
+            return redirect('home');
+        }
         $categories = Category::All();
 
         return view('product/add', ["categories" => $categories]);
@@ -29,6 +38,12 @@ class ProductController extends Controller
 
     public function store(Request $request)
     {
+        $request->validate([
+            'title' => ['required'],
+            'description' => ['required'],
+            'image' => ['required'],
+            'categories' => ['required'],
+        ]);
 
         $product = new Product;
 
@@ -44,9 +59,20 @@ class ProductController extends Controller
 
     public function delete($id)
     {
+        if (!$this->isAdmin()) {
+            return redirect('home');
+        }
+
         $product = Product::find($id);
         $product->delete();
 
         return back()->withInfo('Delete product');
+    }
+
+    public function isAdmin()
+    {
+        $user = Auth::user();
+
+        return (Auth::check() && $user->role === 'admin');
     }
 }
